Improve Contributors Page Structure
