+++ packages/nextjs/components/builder-registry/contributor/ContributorsStats.tsx
