+++ packages/nextjs/components/builder-registry/contributor/index.tsx
@@ -1,0 +1,3 @@
+export * from "./ContributorDetails";
+export * from "./ContributorSkills";
+export * from "./ContributorHeader";
